+++ client/checkout/utils/upe.js
@@ -36,3 +36,31 @@ export const isWCPayChosen = function () {
 	return document.getElementById( 'payment_method_woocommerce_payments' )
 		.checked;
 };
+
+/**
+ * Returns the cached payment intent for the current cart state.
+ *
+ * @param {Object} paymentMethodsConfig Array of configs for payment methods.
+ * @param {string} paymentMethodType Type of the payment method.
+ * @return {Object} The intent id and client secret required for mounting the UPE element.
+ */
+export const getPaymentIntentFromSession = (
+	paymentMethodsConfig,
+	paymentMethodType
+) => {
+	const cartHash = getCookieValue( 'woocommerce_cart_hash' );
+	const upePaymentIntentData =
+		paymentMethodsConfig[ paymentMethodType ].upePaymentIntentData;
+
+	if (
+		cartHash &&
+		upePaymentIntentData &&
+		upePaymentIntentData.startsWith( cartHash )
+	) {
+		const intentId = upePaymentIntentData.split( '-' )[ 1 ];
+		const clientSecret = upePaymentIntentData.split( '-' )[ 2 ];
+		return { intentId, clientSecret };
+	}
+
+	return {};
+};
